expenses-classification/requirements.txt
langchain == 0.1.16
langchain-community == 0.0.33
langchain-core == 0.1.44
langchain-openai == 0.1.3
langchain-text-splitters == 0.0.1
fastapi == 0.110.2
uvicorn == 0.29.0
python-dotenv == 1.0.1
autopep8 == 2.1.0
PyJWT == 2.8.0
psycopg2-binary == 2.9.9